--- a/.travis.yml
+++ b/.travis.yml
@@ -8,6 +8,9 @@ env:
     # may also want to run `go mod edit -go=1.13` to fix go.mod as well
     - GIMME_GO_VERSION=1.13.x GIMME_OS=linux GIMME_ARCH=amd64
 
+services:
+  - redis-server
+
 branches:
   only:
   - master