--- a/cmd/optimizely/main.go
+++ b/cmd/optimizely/main.go
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2019-2021 Optimizely, Inc. and contributors                    *
+ * Copyright 2019,2022, Optimizely, Inc. and contributors                   *
  *                                                                          *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *