--- a/cmd/optimizely/main.go
+++ b/cmd/optimizely/main.go
@@ -87,7 +87,7 @@ func loadConfig(v *viper.Viper) *config.AgentConfig {
 		conf.Server.Interceptors = interceptors
 	}
 
-	// Check if JSON string was set using OPTIMIZELY_CLIENT_USERPROFILESERVICE
+	// Check if JSON string was set using OPTIMIZELY_CLIENT_USERPROFILESERVICE environment variable
 	if userProfileService := v.GetStringMap("client.userprofileservice"); userProfileService != nil {
 		conf.Client.UserProfileService = userProfileService
 	}