--- a/config/config.go
+++ b/config/config.go
@@ -79,18 +79,8 @@ func NewDefaultConfig() *AgentConfig {
 			// https://github.com/google/re2/wiki/Syntax
 			SdkKeyRegex: "^\\w+(:\\w+)?$",
 			UserProfileServices: UserProfileServiceConfigs{
-				"default": "in-memory",
-				"services": map[string]interface{}{
-					"in-memory": map[string]interface{}{
-						"capacity": 0,
-					},
-					"redis": map[string]interface {
-					}{
-						"host":     "localhost:6379",
-						"password": "",
-						"database": 0,
-					},
-				},
+				"default":  "",
+				"services": map[string]interface{}{},
 			},
 		},
 		Runtime: RuntimeConfig{