--- a/go.mod
+++ b/go.mod
@@ -10,6 +10,7 @@ require (
 	github.com/go-chi/httplog v0.1.6
 	github.com/go-chi/render v1.0.1
 	github.com/go-kit/kit v0.9.0
+	github.com/go-redis/redis/v8 v8.11.4
 	github.com/google/uuid v1.1.1
 	github.com/lestrrat-go/jwx v0.9.0
 	github.com/optimizely/go-sdk v1.7.1-0.20211102215133-82644c6670e8
@@ -17,11 +18,10 @@ require (
 	github.com/rakyll/statik v0.1.7
 	github.com/rs/zerolog v1.18.1-0.20200514152719-663cbb4c8469
 	github.com/spf13/viper v1.4.0
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/crypto v0.0.0-20200221231518-2aa609cf4a9d
-	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e
-	golang.org/x/sys v0.0.0-20191105231009-c1f44814a5cd // indirect
-	gopkg.in/yaml.v2 v2.2.4
+	github.com/stretchr/testify v1.5.1
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
+	golang.org/x/sync v0.0.0-20201020160332-67f06af15bc9
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 exclude github.com/gorilla/websocket v1.4.0