… README.md
… api/openapi-spec/openapi.yaml
… cmd/optimizely/main_test.go
… cmd/optimizely/testdata/default.yaml
… config.yaml
… config/config_test.go
… go.sum
… pkg/handlers/lookup_test.go
… pkg/handlers/save_test.go
… pkg/handlers/webhook_test.go
… pkg/middleware/cached_test.go
… pkg/optimizely/cache_test.go
… pkg/optimizely/client.go
… pkg/optimizely/interface.go
… pkg/optimizely/user_profile_service.go
… pkg/routers/api.go
… pkg/routers/api_test.go
… plugins/interceptors/registry_test.go
… plugins/userprofileservice/README.md
… plugins/userprofileservice/all/all.go
… plugins/userprofileservice/memory/in_memory_ups.go
… plugins/userprofileservice/registry.go
… plugins/userprofileservice/registry_test.go
… plugins/userprofileservice/services/in_memory_ups.go
… plugins/userprofileservice/services/in_memory_ups_test.go
… plugins/userprofileservice/services/redis_ups.go
… plugins/userprofileservice/services/redis_ups_test.go
… plugins/userprofileservice/services/rest_ups.go
… plugins/userprofileservice/services/rest_ups_test.go
… plugins/userprofileservice/services/utils.go