@yasirfolio3 2022-01-07
> Hey team, this is amazing news and something my team was considering contributing! 💟
> 
> Just a question, in the example config we see there is an option to specify custom URL for User Profile Service. However, reviewing PR I could not see an implementation for that. Is there a plan to support that any time soon? Even the PR description does not cover this.
> 
> Thank you :-)

Hi @anvth , We have updated the PR description, please have a look and let us know if there's anything else we can help with, Thanks.