@204Constie 2022-01-10
Hey team, this is amazing news for us as we were planning to implement this. 😸 
A significant question here that we would have would be about the custom UPS implementation. Contrary to how in memory and redis are implemented there is no way of implementing it by simply providing certain options in the config file. 
Did you consider an option to make it less flexible but not requiring changes in the code ? 
What about having an out of the box option for microservice UPS ?

thank you 😸