@yasirfolio3 2022-01-28
> > > hello! 😸 all of the updates to rest UPS look wonderful 😻 all configurable via config.yml file 😻 great work! i have a small question. i can read that both `lookup` and `save` methods expect the rest endpoints to return the profile on `success`. this is for sure necessary in case of `lookup` but is it really required in case of `save` method ? does it use the response of the `save` method or expect to use it in the future ? or it actually only require the response of status code `2XX` and would in truth behave the same in case of `200` and `204` ?
> > 
> > 
> > Hi @204Constie , In case of `save`, only 200 status code needs to be sent if successful (204 is not supported). The documentation is actually stating the payload required for the `save` request 👍
> 
> 😸 this is perfectly fine, i was just wondering why is this the case. is the response used in any way ? is there some other reason ? i'm simply curious

So for `save` API, no response is required. Even if a response is sent, it will be ignored by agent. Agent only checks for status code `200`.